refactor(slack): consolidate test helpers and remove unit tests
- Remove internal handler unit tests (mention.test.ts, run-agent.test.ts)
  in favor of API route integration tests
- Move test helpers from lib/slack/__tests__/helpers to src/__tests__/slack
- Update all Slack test files to use new helper location
- Make helper types internal (not exported) to satisfy knip
