+++ turbo/apps/web/app/api/slack/oauth/callback/__tests__/route.test.ts
@@ -1,4 +1,4 @@
-import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
+import { describe, it, expect, vi, beforeEach } from "vitest";
 import { http, HttpResponse } from "msw";
 import { GET } from "../route";
 import { createTestRequest } from "../../../../../../src/__tests__/api-test-helpers";
@@ -18,34 +18,9 @@ const context = testContext();
 describe("/api/slack/oauth/callback", () => {
   beforeEach(() => {
     context.setupMocks();
-    vi.clearAllMocks();
-    // Reset to default test values
-    vi.stubEnv("SLACK_CLIENT_ID", "test-slack-client-id");
-    vi.stubEnv("SLACK_CLIENT_SECRET", "test-slack-client-secret");
-    vi.stubEnv("SLACK_REDIRECT_BASE_URL", "");
-    reloadEnv();
-  });
-
-  afterEach(() => {
-    server.resetHandlers();
   });
 
   describe("GET /api/slack/oauth/callback", () => {
-    it("should return 503 when Slack credentials are not configured", async () => {
-      vi.stubEnv("SLACK_CLIENT_ID", "");
-      vi.stubEnv("SLACK_CLIENT_SECRET", "");
-      reloadEnv();
-
-      const request = createTestRequest(
-        "http://localhost:3000/api/slack/oauth/callback?code=test-code",
-      );
-      const response = await GET(request);
-      const data = await response.json();
-
-      expect(response.status).toBe(503);
-      expect(data.error).toBe("Slack integration is not configured");
-    });
-
     it("should redirect to failed page when error parameter is present", async () => {
       const request = createTestRequest(
         "http://localhost:3000/api/slack/oauth/callback?error=access_denied",
@@ -55,7 +30,7 @@ describe("/api/slack/oauth/callback", () => {
       expect(response.status).toBe(307);
       const locationHeader = response.headers.get("Location");
       expect(locationHeader).toBe(
-        "http://localhost:3000/slack/failed?error=access_denied",
+        "https://test.example.com/slack/failed?error=access_denied",
       );
     });
