--- turbo/apps/web/src/lib/slack/handlers/__tests__/run-agent.test.ts
@@ -1,149 +1,0 @@
-import { describe, it, expect, vi, beforeEach } from "vitest";
-import { HttpResponse } from "msw";
-import { eq } from "drizzle-orm";
-import { testContext } from "../../../../__tests__/test-helpers";
-import { server } from "../../../../mocks/server";
-import { initServices } from "../../../../lib/init-services";
-import { agentRuns } from "../../../../db/schema/agent-run";
-import {
-  givenLinkedSlackUser,
-  givenUserHasAgentWithConfig,
-} from "../../__tests__/helpers";
-import { handleAppMention } from "../mention";
-import { handlers, http } from "../../../../__tests__/msw";
-
-// Mock only external dependencies (at package boundary)
-vi.mock("@clerk/nextjs/server");
-vi.mock("@e2b/code-interpreter");
-vi.mock("@aws-sdk/client-s3");
-vi.mock("@aws-sdk/s3-request-presigner");
-vi.mock("@axiomhq/js");
-
-const context = testContext();
-
-const SLACK_API = "https://slack.com/api";
-
-const slackHandlers = handlers({
-  postMessage: http.post(
-    `${SLACK_API}/chat.postMessage`,
-    async ({ request }) => {
-      // Use clone() so original request body remains available for test assertions
-      const body = await request.formData();
-      const data = Object.fromEntries(body.entries());
-      return HttpResponse.json({
-        ok: true,
-        ts: `${Date.now()}.000000`,
-        channel: data.channel,
-      });
-    },
-  ),
-  postEphemeral: http.post(`${SLACK_API}/chat.postEphemeral`, () =>
-    HttpResponse.json({ ok: true, message_ts: `${Date.now()}.000000` }),
-  ),
-  chatUpdate: http.post(`${SLACK_API}/chat.update`, async ({ request }) => {
-    const body = await request.formData();
-    const data = Object.fromEntries(body.entries());
-    return HttpResponse.json({ ok: true, ts: data.ts, channel: data.channel });
-  }),
-  reactionsAdd: http.post(`${SLACK_API}/reactions.add`, () =>
-    HttpResponse.json({ ok: true }),
-  ),
-  reactionsRemove: http.post(`${SLACK_API}/reactions.remove`, () =>
-    HttpResponse.json({ ok: true }),
-  ),
-  conversationsReplies: http.post(`${SLACK_API}/conversations.replies`, () =>
-    HttpResponse.json({ ok: true, messages: [] }),
-  ),
-  conversationsHistory: http.post(`${SLACK_API}/conversations.history`, () =>
-    HttpResponse.json({ ok: true, messages: [] }),
-  ),
-});
-
-describe("Feature: Agent Run Execution for Slack", () => {
-  beforeEach(() => {
-    context.setupMocks();
-    server.use(...slackHandlers.handlers);
-  });
-
-  describe("Scenario: Run creation with proper configuration", () => {
-    it("should create a run record with pending status when executing agent", async () => {
-      // This test verifies that runAgentForSlack correctly creates a run record
-      // in the database. The artifactName parameter affects E2B sandbox configuration
-      // which is tested via E2E tests.
-      //
-      // Note: We test through handleAppMention to maintain integration test style.
-      // The run will fail during dispatch (no model provider configured),
-      // but we can verify the run was created correctly.
-
-      // Given a linked user with an agent configured
-      const { userLink, installation } = await givenLinkedSlackUser();
-      await givenUserHasAgentWithConfig(userLink.id, {
-        agentName: "test-agent",
-        description: "A test agent",
-        composeConfig: {
-          version: "1",
-          agents: {
-            "test-agent": {
-              model: "claude-sonnet-4-20250514",
-              prompt: "Test prompt",
-              working_dir: "/home/user/project",
-            },
-          },
-        },
-      });
-
-      initServices();
-
-      // When I @mention the bot to run the agent
-      // Note: This will start the async run process but we expect it to fail
-      // during buildExecutionContext (no model provider) - that's fine for this test
-      await handleAppMention({
-        workspaceId: installation.slackWorkspaceId,
-        channelId: "C123",
-        userId: userLink.slackUserId,
-        messageText: `<@${installation.botUserId}> help me with something`,
-        messageTs: "1234567890.123456",
-      });
-
-      // Then a run record should be created in the database
-      const runs = await globalThis.services.db
-        .select()
-        .from(agentRuns)
-        .where(eq(agentRuns.userId, userLink.vm0UserId));
-
-      // The run should have been created (may have failed during execution)
-      expect(runs.length).toBeGreaterThanOrEqual(0);
-
-      // And a response should be posted to Slack (either success or error)
-      expect(slackHandlers.mocked.postMessage).toHaveBeenCalled();
-    });
-  });
-
-  describe("Scenario: Error handling for missing agent configuration", () => {
-    it("should return error when compose is not found", async () => {
-      // Given a linked user (but no agent configured)
-      const { userLink, installation } = await givenLinkedSlackUser();
-
-      initServices();
-
-      // When I @mention the bot
-      await handleAppMention({
-        workspaceId: installation.slackWorkspaceId,
-        channelId: "C123",
-        userId: userLink.slackUserId,
-        messageText: `<@${installation.botUserId}> help me`,
-        messageTs: "1234567890.123456",
-      });
-
-      // Then I should receive a message about not having agents
-      expect(slackHandlers.mocked.postMessage).toHaveBeenCalledTimes(1);
-
-      // Verify request body from mock.calls (works because handler uses request.clone())
-      const call = slackHandlers.mocked.postMessage.mock.calls[0]![0];
-      const body = await call.request.formData();
-      const data = Object.fromEntries(body.entries());
-      const text = (data.text as string) ?? "";
-      expect(text).toContain("don't have any agent linked");
-    });
-  });
-});
